feat: Add bulk import, search/filter, and fix session update bug
- Add bulk question import with CSV/JSON support and XLSX template
- Add search and filter to QuestionList (with pagination) and RoundEditor
- Add search to PlayerSelector in session builder
- Add color-coded badges for question types (QCM=blue, Closest=purple, Open=green, Image=yellow)
- Add notes field to Question model and display in host view
- Fix closest question float value support (z.number() instead of z.number().int())
- Fix critical session update bug where edited sessions weren't saved to disk
- Hide already-added questions/guests from selection lists
- Add duplicate detection for bulk imports with visual indicators
- Improve UI layout with adjacent action buttons and better spacing

+++ lib/models/Quiz.ts
@@ -36,7 +36,7 @@ export const questionSchema = z.object({
   text: z.string(),
   media: z.string().nullable().default(null), // Allow both URLs and relative paths
   options: z.array(z.string()).max(4).optional(),
-  correct: z.union([z.number().int(), z.string(), z.object({ min: z.number(), max: z.number() })]).optional(),
+  correct: z.union([z.number(), z.string(), z.object({ min: z.number(), max: z.number() })]).optional(),
   points: z.number().int().default(1),
   tie_break: z.boolean().default(false),
   time_s: z.number().int().positive().default(20),
